+++ src/util/sexpr_file_reader.cpp
@@ -107,7 +107,7 @@ public:
           }
         else if (item->get_type() == lisp::Lisp::TYPE_INT)
           {
-            v = item->get_int();
+            v = static_cast<float>(item->get_int());
             return true;
           }
         else
